+++ CHANGELOG.md
@@ -2,8 +2,9 @@
 
 * Add support for turning off the heater entirely. The last preferred target temperature is persisted to local device storage.
   Fixes issue #1.
-* Fix the liquid level computation.
 * Use a fancier, graphical liquid level indicator, instead of a text. 
+* Fix the liquid level computation, as per issue #3.
+* Add a state icon, as per issue #3.
 
 # 2023-02-16